[scout] split tests, fix tags (#263850)
## Summary

Splitting tests and adjusting tags based on Cloud failures

Closes  https://github.com/elastic/kibana/issues/263236
Closes https://github.com/elastic/kibana/issues/263237

---------

(cherry picked from commit fef4bb74246a8b9a4e743e9899bef6d3c9ae36ae)
